requirements.txt
requests>=2.31.0
requests-oauthlib>=1.3.1
schedule>=1.2.0
google-generativeai>=0.8.3
